refactor: add minor improvements in code
