+++ app/components/ExploreHeader.scss
